TestApplication/requirements.txt
Flask==1.0.2
google-cloud-datastore
flask-login
google-cloud-ndb
googleapis_common_protos